Rename local connection files and dev commands
