+++ cjs/src/server.local.js
@@ -1,5 +1,5 @@
-import mongooseConnect from './db.connect';
-import { app as app } from './app';
+const mongooseConnect = require('./db.local.connect');
+const { app } = require('./app');
 
 const port = process.env.PORT || 3000;
