+++ ts/package.json
@@ -4,7 +4,7 @@
   "description": "@code-collabo node-mongo API boilerplate template - typescript",
   "main": "src/server.[atlas/local].ts",
   "scripts": {
-    "dev": "ts-node-dev --respawn --transpile-only src/server.ts",
+    "dev:local": "ts-node-dev --respawn --transpile-only src/server.local.ts",
     "dev:atlas": "ts-node-dev --respawn --transpile-only src/server.atlas.ts",
     "lint": "esw src --ext .ts --color",
     "lint:watch": "npm run lint -- --watch"
