+++ ts/src/server.local.ts
@@ -1,5 +1,5 @@
-const mongooseConnect = require('./db.connect');
-const { app } = require('./app');
+import mongooseConnect from './db.local.connect';
+import { app as app } from './app';
 
 const port = process.env.PORT || 3000;
